requirements.txt
fastapi[standard]
uvicorn 
sqlalchemy 
pymysql 
python-dotenv 
bcrypt 
passlib
passlib[argon2]
python-jose

